Working towards 2.0.1: Fix NPE bug when reset-on-startup is on and no data exists

+++ src/main/java/com/bergerkiller/bukkit/mw/WorldConfig.java
@@ -1753,16 +1753,18 @@ public class WorldConfig extends WorldConfigStore {
         // Delete the region, poi and entities folder entirely. This always exists regardless of format.
         {
             File regionFolder = loadableWorld.getRegionFolder();
-            if (regionFolder.exists()) {
-                if (!StreamUtil.deleteFile(regionFolder).isEmpty()) {
-                    fullySuccessful = false;
+            if (regionFolder != null) {
+                if (regionFolder.exists()) {
+                    if (!StreamUtil.deleteFile(regionFolder).isEmpty()) {
+                        fullySuccessful = false;
+                    }
                 }
-            }
-            File parentFolder = regionFolder.getParentFile();
-            for (String otherFolderName : new String[] {"poi", "entities"}) {
-                File otherFolder = new File(parentFolder, otherFolderName);
-                if (otherFolder.exists() && !StreamUtil.deleteFile(otherFolder).isEmpty()) {
-                    fullySuccessful = false;
+                File parentFolder = regionFolder.getParentFile();
+                for (String otherFolderName : new String[] {"poi", "entities"}) {
+                    File otherFolder = new File(parentFolder, otherFolderName);
+                    if (otherFolder.exists() && !StreamUtil.deleteFile(otherFolder).isEmpty()) {
+                        fullySuccessful = false;
+                    }
                 }
             }
         }
